% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLikelihood.R
\name{sLikelihood}
\alias{sLikelihood}
\title{Symbolic Pedigree Likelihood}
\usage{
sLikelihood(x)
}
\arguments{
\item{x}{A \code{ped} object.}
}
\value{
A \code{character} with the symbolic likelihood
}
\description{
Symbolic Pedigree Likelihood
}
\examples{
# likelihood for a single heterozygous genotype
require(pedtools)
het <- singleton() |> addMarker(geno="a/b")
sLikelihood(het) # 2*a*b

# full siblings
fs <- nuclearPed(nch = 2) |> addMarker(geno = c(NA, NA, "a/a", "a/a"))
sLikelihood(fs) # "0.25*a^4+0.5*a^3+0.25*a^2"
}
